pyproject.toml
[tool.poetry]
name = "repl-nix-diabeticretinopathyanalyzer-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
openai = "^1.50.2"
pandas = "^2.2.3"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
